# api.yaml
openapi: 3.0.0
info:
  title: "Szabadságnyilvántartó API"
  version: 1.0.0
paths:
...
  /tappenz/start:
    {
      post:
        {
          summary: "Start a new tappenz entry with only starting date filled out",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          properties:
                            {
                              userId:
                                {
                                  type: string,
                                  description: "The ID of the user starting the tappenz",
                                },
                              start:
                                {
                                  type: string,
                                  format: date,
                                  description: "The start date of the tappenz in format YYYY-MM-DD",
                                },
                            },
                        },
                    },
                },
            },
          responses:
            {
              "200":
                {
                  description: "Returns a status and a tappenz document or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  tappenz:
                                    { $ref: "#/components/schemas/Tappenz" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  "/tappenz/current/{id}":
    {
      get:
        {
          summary: "Retrieve the current tappenz status of a user by ID",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the user whose current tappenz status to retrieve",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a boolean indicating if the user is currently in tappenz or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  current:
                                    {
                                      type: boolean,
                                      description: "Whether the user is currently in tappenz",
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  /tappenz/end:
    {
      post:
        {
          summary: "End a tappenz entry by filling out the end date",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          properties:
                            {
                              userId:
                                {
                                  type: string,
                                  description: "The ID of the user ending the tappenz",
                                },
                              end:
                                {
                                  type: string,
                                  format: date,
                                  description: "The end date of the tappenz in format YYYY-MM-DD",
                                },
                            },
                        },
                    },
                },
            },
          responses:
            {
              "200":
                {
                  description: "Returns a status and a tappenz document or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  tappenzDoc:
                                    { $ref: "#/components/schemas/Tappenz" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  "/tappenz/{id}":
    {
      delete:
        {
          summary: "Delete a tappenz entry by ID",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the tappenz entry to delete",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a tappenz document or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  tappenzDoc:
                                    { $ref: "#/components/schemas/Tappenz" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
      get:
        {
          summary: "Retrieve the last 5 tappenz entries of a user by ID",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the user whose last 5 tappenz entries to retrieve",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a list of tappenz entries or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  tappenz:
                                    {
                                      type: array,
                                      items:
                                        {
                                          $ref: "#/components/schemas/Tappenz",
                                        },
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  "/tappenz/{id}/cumulative":
    {
      get:
        {
          summary: "Retrieve the cumulative number of days a user has taken as sick leave by ID",
          tags: [Tappenz],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the user whose cumulative sick leave days to retrieve",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and the cumulative number of sick leave days or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  cumulative:
                                    {
                                      type: integer,
                                      description: "The cumulative number of sick leave days",
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
...
components:
  securitySchemes:
    { ApiKeyAuth: { type: apiKey, in: header, name: submittingId } }
  schemas:
...
      Tappenz:
        {
          type: object,
          properties:
            {
              startDate: { format: date },
              endDate: { format: date },
              userId: { type: string },
              managerId: { type: string },
            },
        },
...